+++ src/domains/User/controllers/index.ts
@@ -1,0 +1,6 @@
+import { Router, Request, Response, NextFunction } from "express";
+import serviceUsuario from "../services/serviceUsuario";
+
+const router = Router();
+
+export default router;
